Quiet compiler warnings.

+++ src/backend/postmaster/postmaster.c
@@ -10,7 +10,7 @@
  *
  *
  * IDENTIFICATION
- *    $Header: /cvsroot/pgsql/src/backend/postmaster/postmaster.c,v 1.21 1996/11/08 05:57:41 momjian Exp $
+ *    $Header: /cvsroot/pgsql/src/backend/postmaster/postmaster.c,v 1.22 1996/11/10 02:24:12 bryanh Exp $
  *
  * NOTES
  *
@@ -1170,8 +1170,8 @@ dumpstatus(SIGNAL_ARGS)
 	fprintf(stderr, "%s: dumpstatus:\n", progname);
 	fprintf(stderr, "\tsock %d: nBytes=%d, laddr=0x%lx, raddr=0x%lx\n",
 		port->sock, port->nBytes, 
-		port->laddr.sin_addr.s_addr, 
-		port->raddr.sin_addr.s_addr);
+		(long int) port->laddr.sin_addr.s_addr, 
+		(long int) port->raddr.sin_addr.s_addr);
 	curr = DLGetSucc(curr);
     }
 }
